requirements.txt
numpy
pandas
opencv-python
cherrypy
configparser
adafruit-circuitpython-neopixel
blinker
jinja2
sounddevice